More changes based on code review
