+++ src/main/java/org/kiwiproject/dropwizard/util/exception/EarlyEofExceptionMapper.java
@@ -3,6 +3,7 @@ package org.kiwiproject.dropwizard.util.exception;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.jetty.io.EofException;
 import org.kiwiproject.jaxrs.exception.JaxrsBadRequestException;
+import org.kiwiproject.jaxrs.exception.JaxrsException;
 import org.kiwiproject.jaxrs.exception.JaxrsExceptionMapper;
 
 import javax.ws.rs.core.Response;
