+++ src/main/java/org/kiwiproject/dropwizard/util/exception/EmptyOptionalExceptionMapper.java
@@ -1,6 +1,7 @@
 package org.kiwiproject.dropwizard.util.exception;
 
 import io.dropwizard.jersey.optional.EmptyOptionalException;
+import org.kiwiproject.jaxrs.exception.JaxrsException;
 import org.kiwiproject.jaxrs.exception.JaxrsExceptionMapper;
 import org.kiwiproject.jaxrs.exception.JaxrsNotFoundException;
 
@@ -9,7 +10,8 @@ import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
 
 /**
- * Override default Dropwizard mapper to use common error structure.
+ * Override default Dropwizard mapper to use kiwi's {@link org.kiwiproject.jaxrs.exception.ErrorMessage ErrorMessage}.
+ * The the response entity is built using {@link JaxrsExceptionMapper#buildResponseEntity(JaxrsException)}.
  */
 @Provider
 public class EmptyOptionalExceptionMapper implements ExceptionMapper<EmptyOptionalException> {
