+++ src/test/java/org/kiwiproject/dropwizard/util/exception/EarlyEofExceptionMapperTest.java
@@ -1,6 +1,8 @@
 package org.kiwiproject.dropwizard.util.exception;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.kiwiproject.dropwizard.util.exception.ErrorMessageAssertion.assertAndGetErrorMessage;
+import static org.kiwiproject.test.jaxrs.JaxrsTestHelper.assertBadRequest;
 
 import org.eclipse.jetty.io.EofException;
 import org.junit.jupiter.api.DisplayName;
@@ -13,6 +15,10 @@ class EarlyEofExceptionMapperTest {
     void testMapper() {
         var mapper = new EarlyEofExceptionMapper();
         var response = mapper.toResponse(new EofException("oops"));
-        assertThat(response.getStatus()).isEqualTo(400);
+
+        assertBadRequest(response);
+
+        var errorMessage = assertAndGetErrorMessage(response);
+        assertThat(errorMessage.getMessage()).isEqualTo("oops");
     }
 }
