+++ src/test/java/org/kiwiproject/dropwizard/util/exception/EmptyOptionalExceptionMapperTest.java
@@ -1,5 +1,7 @@
 package org.kiwiproject.dropwizard.util.exception;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.kiwiproject.dropwizard.util.exception.ErrorMessageAssertion.assertAndGetErrorMessage;
 import static org.kiwiproject.test.jaxrs.JaxrsTestHelper.assertNotFoundResponse;
 
 import io.dropwizard.jersey.optional.EmptyOptionalException;
@@ -13,6 +15,10 @@ class EmptyOptionalExceptionMapperTest {
     void shouldReturnNotFoundResponse() {
         var ex = EmptyOptionalException.INSTANCE;
         var response = new EmptyOptionalExceptionMapper().toResponse(ex);
+
         assertNotFoundResponse(response);
+
+        var errorMessage = assertAndGetErrorMessage(response);
+        assertThat(errorMessage.getMessage()).isEqualTo("Unknown error");
     }
 }
